Renaming part 2.
