+++ src/components/App/App.js
@@ -1,9 +1,9 @@
-import Header from '../Aheader/Header'
-import Main from '../Amain/Main'
-import Footer from '../Afooter/Footer'
-import Home from '../Ahome/Home'
+import Header from '../Header/Header'
+import Main from '../Main/Main'
+import Footer from '../Footer/Footer'
+import Home from '../Home/Home'
 import ActivitiesList from '../ActivitiesList/ActivitiesList'
-import CreateActivityForm from '../AcreateActivityForm/CreateActivityForm'
+import CreateActivityForm from '../CreateActivityForm/CreateActivityForm'
 import Map from '../Map/Map'
 import loadFromLocal from '../../lib/loadFromLocal'
 import saveToLocal from '../../lib/saveToLocal'
