+++ src/components/BookmarkButton/BookmarkButton.spec.js
