+++ src/components/FlipCard/FlipCard.spec.js
