+++ src/components/Footer/Footer.spec.js
