+++ src/components/SearchBox/SearchBox.spec.js
