+++ src/components/SearchBox/SearchBox.stories.js
